move all in silico predictors functions to a new file under annotations
